Se realizo un soporte express para cuando sea haga el cambio de puesto del integrante en una originacion grupal o renovacion grupal aun conserve los beneficiarios y la campaña de la solicitud, esto porque se reutiliza la vista de registro de integrante, cuando se crea un nuevo integrante y se cambia el rol del integrante en el grupo se utiliza la misma vista, solo que con diferente configuracion, recarga solo el nombre, y algunos datos pero ahora tambien recarga la informacion del beneficiario y de la campaña.

+++ app/src/main/java/com/sidert/sidertmovil/database/dao/BeneficiariosDao.java
@@ -17,6 +17,6 @@ public interface BeneficiariosDao {
 
     void delete(Beneficiario entity);
 
-    Optional<Beneficiario> findBySolicitudId(Long solicitudId, Integer integranteId, String tipoSolicitud);
+    Optional<Beneficiario> findBySolicitudId(Integer solicitudId, Integer integranteId, String tipoSolicitud);
 
 }
